Fix issues with typesafe accessor behavior in strict mode (#37)
Plugin was supplying mapOf() to the rule when STRICT mode (the default) was used, which actually implies FULL mode to the buildscript-utils code, and then never matches any type-safe accessors because there is nothing to match. The value passed should have been null.

Solved this by splitting the rule in two, one class for each mode.